requirements.txt
numpy
numpy-stl
seaborn
matplotlib
scipy
scikit-learn